+++ src/serial/3.c
@@ -2,14 +2,14 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
-#include "../lib/headers/generator.h"
+#include "../../lib/headers/generator.h"
 
 int main(int argc, char const *argv[])
 {
-    float *v = (float *)malloc(SIZE * sizeof(float));
+    double *v = (double *)malloc(SIZE * sizeof(double));
     gen_vector(v);
 
-    float average = 0;
+    double average = 0;
 
     double t_start = omp_get_wtime();
